+++ lib/devops/lib/model/github-trigger.ts
@@ -22,6 +22,10 @@ export interface GithubTrigger extends model.Trigger {
    * The endpoint that listens to trigger events.
    */
   "triggerUrl": string;
+  /**
+   * The OCID of the connection resource used to get details for triggered events.
+   */
+  "connectionId"?: string;
 
   "triggerSource": string;
 }