+++ lib/devops/lib/model/gitlab-filter-attributes.ts
@@ -26,16 +26,29 @@ export interface GitlabFilterAttributes {
    * The target branch for pull requests; not applicable for push requests.
    */
   "baseRef"?: string;
+  "fileFilter"?: model.FileFilter;
 }
 
 export namespace GitlabFilterAttributes {
   export function getJsonObj(obj: GitlabFilterAttributes): object {
-    const jsonObj = { ...obj, ...{} };
+    const jsonObj = {
+      ...obj,
+      ...{
+        "fileFilter": obj.fileFilter ? model.FileFilter.getJsonObj(obj.fileFilter) : undefined
+      }
+    };
 
     return jsonObj;
   }
   export function getDeserializedJsonObj(obj: GitlabFilterAttributes): object {
-    const jsonObj = { ...obj, ...{} };
+    const jsonObj = {
+      ...obj,
+      ...{
+        "fileFilter": obj.fileFilter
+          ? model.FileFilter.getDeserializedJsonObj(obj.fileFilter)
+          : undefined
+      }
+    };
 
     return jsonObj;
   }