+++ lib/devops/lib/model/update-gitlab-trigger-details.ts
@@ -18,6 +18,11 @@ import common = require("oci-common");
  * Update trigger specific to GitLab.
  */
 export interface UpdateGitlabTriggerDetails extends model.UpdateTriggerDetails {
+  /**
+   * The OCID of the connection resource used to get details for triggered events.
+   */
+  "connectionId"?: string;
+
   "triggerSource": string;
 }