+++ lib/mediaservices/lib/model/create-media-workflow-details.ts
@@ -48,7 +48,7 @@ export interface CreateMediaWorkflowDetails {
    * MediaWorkflowJobs from this MediaWorkflow.
    *
    */
-  "parameters"?: { [key: string]: string };
+  "parameters"?: { [key: string]: any };
   /**
    * Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.
    * Example: `{\"bar-key\": \"value\"}`