+++ lib/optimizer/lib/request/list-resource-actions-request.ts
@@ -32,7 +32,43 @@ Can only be set to true when performing ListCompartments on the tenancy (root co
   /**
    * The unique OCID associated with the recommendation.
    */
-  "recommendationId": string;
+  "recommendationId"?: string;
+  /**
+   * Optional. A filter that returns results that match the recommendation name specified.
+   */
+  "recommendationName"?: string;
+  /**
+ * A list of child tenancies for which the respective data will be returned. Please note that 
+* the parent tenancy id can also be included in this list. For example, if there is a parent P with two
+* children A and B, to return results of only parent P and child A, this list should be populated with 
+* tenancy id of parent P and child A. 
+* <p>
+If this list contains a tenancy id that isn't part of the organization of parent P, the request will 
+* fail. That is, let's say there is an organization with parent P with children A and B, and also one 
+* other tenant T that isn't part of the organization. If T is included in the list of 
+* childTenancyIds, the request will fail.
+* <p>
+It is important to note that if you are setting the includeOrganization parameter value as true and 
+* also populating the childTenancyIds parameter with a list of child tenancies, the request will fail.
+* The childTenancyIds and includeOrganization should be used exclusively.
+* <p>
+When using this parameter, please make sure to set the compartmentId with the parent tenancy ID.
+* 
+ */
+  "childTenancyIds"?: Array<string>;
+  /**
+ * When set to true, the data for all child tenancies including the parent is returned. That is, if 
+* there is an organization with parent P and children A and B, to return the data for the parent P, child 
+* A and child B, this parameter value should be set to true.
+* <p>
+Please note that this parameter shouldn't be used along with childTenancyIds parameter. If you would like 
+* to get results specifically for parent P and only child A, use the childTenancyIds parameter and populate
+* the list with tenancy id of P and A.
+* <p>
+When using this parameter, please make sure to set the compartmentId with the parent tenancy ID.
+* 
+ */
+  "includeOrganization"?: boolean;
   /**
    * Optional. A filter that returns results that match the name specified.
    */